Add authentication, /places endpoint, use secrets for sensitive info

+++ api/controllers/users.py
@@ -1,7 +1,10 @@
 from flask import Blueprint, jsonify, request
 
-from ..models import User, db
+from ..auth import admin_only, authenticate, get_current_user
+from ..exceptions import ApiException
+from ..models import Plan, User, db
 from ..schemas import UserSchema
+from .plans import plans_schema
 
 bp = Blueprint("users", __name__)
 
@@ -22,9 +25,27 @@ def get_user(id):
     return jsonify(user_schema.dump(user)), 200
 
 
+@bp.route("/<int:user_id>/plans", methods=["GET"])
+def get_users_plans(user_id):
+    plans = User.query.get_or_404(user_id).plans
+    return plans_schema.dump(plans)
+
+
+@bp.route("/<int:user_id>/plans/<int:plan_id>", methods=["GET"])
+def get_users_plan(user_id, plan_id):
+    plan = Plan.get_or_404(plan_id)
+    if plan.user_id != user_id:
+        raise ApiException("Resource not found", 404)
+
+
 @bp.route("/<int:id>", methods=["POST", "PUT", "PATCH"])
+@authenticate
 def update_user(id):
     user = User.query.get_or_404(id)
+
+    if not user.belongs_to(get_current_user()):
+        raise ApiException("You are not authorized to edit this resource.", 403)
+
     data = request.get_json()
     user.update(**data)
     db.session.commit()
@@ -33,15 +54,21 @@ def update_user(id):
 
 
 @bp.route("/<int:id>", methods=["DELETE"])
+@admin_only
 def delete_user(id):
     user = User.query.get_or_404(id)
+
+    if user.id == get_current_user().id:
+        return ApiException("A user cannot delete itself.", 400)
+
     db.session.delete(user)
     db.session.commit()
 
-    return "", 200
+    return "", 204
 
 
 @bp.route("/", methods=["POST"])
+@admin_only
 def new_user():
     user_data = request.get_json()
     user = user_schema.load(user_data)
